RAG-back/requirements.txt
# Core dependencies - Sistema RAG con Ollama y Gemini

# LLM Providers
ollama==0.3.3
google-generativeai==0.8.3

# Vector Store
chromadb==1.5.7

# Document Processing
pypdf==5.1.0
python-docx==1.1.2

# Utilities
python-dotenv==1.0.1
pyyaml==6.0.2
requests==2.32.3

# API
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.10.3
python-multipart==0.0.20